requirements.txt
flask
flask-sqlalchemy
fpdf2
